system testing tweaks
changed can_receive() return type to 32bit to meet frame length
added can rx handling for 0x314: bms voltage
prevented precharge from restarting instantly after failing, now
requires toggling e-stop

+++ source/vcu.h
@@ -65,7 +65,7 @@ typedef struct {
     uint8_t TS_RDY;             // GPIO E2
     uint8_t TS_LIVE;            // GPIO B7
     uint8_t CHARGER_CONNECTED;  // CAN 0
-    int16_t BMS_VOLTAGE;        // CAN 0
+    int32_t BMS_VOLTAGE;        // CAN 0
     int16_t BMS_TEMPERATURE;    // CAN 0
     uint8_t BMS_OK;             // GPIO E3
     uint8_t IMD_OK;             // GPIO D16
